requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.6
djangorestframework==3.8.2
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.4
static3==0.7.0